+++ apps/io-func-sign-user/src/infra/http/handlers/__test__/get-signature-request.spec.ts
@@ -17,7 +17,6 @@ import {
 } from "../../../../signature-request";
 
 import { GetSignatureRequestHandler } from "../get-signature-request";
-import { BaseContainerClientWithFallback } from "@pagopa/azure-storage-migration-kit";
 import { ContainerClient } from "@azure/storage-blob";
 
 describe("GetSignatureRequestHandler", () => {
@@ -89,7 +88,7 @@ describe("GetSignatureRequestHandler", () => {
       inputDecoder: H.HttpRequest,
       input: req,
       validatedContainerClient: {} as ContainerClient,
-      signedContainerClient: {} as BaseContainerClientWithFallback,
+      signedContainerClient: {} as ContainerClient,
     });
     await expect(run()).resolves.toEqual(
       expect.objectContaining({
@@ -120,7 +119,7 @@ describe("GetSignatureRequestHandler", () => {
       inputDecoder: H.HttpRequest,
       input: req,
       validatedContainerClient: {} as ContainerClient,
-      signedContainerClient: {} as BaseContainerClientWithFallback,
+      signedContainerClient: {} as ContainerClient,
     });
     await expect(run()).resolves.toEqual(
       expect.objectContaining({
@@ -151,7 +150,7 @@ describe("GetSignatureRequestHandler", () => {
       inputDecoder: H.HttpRequest,
       input: req,
       validatedContainerClient: {} as ContainerClient,
-      signedContainerClient: {} as BaseContainerClientWithFallback,
+      signedContainerClient: {} as ContainerClient,
     });
     await expect(run()).resolves.toEqual(
       expect.objectContaining({