+++ apps/io-func-sign-user/CHANGELOG.md
@@ -1,5 +1,17 @@
 # io-func-sign-user
 
+## 2.9.5
+
+### Patch Changes
+
+- 4f0a261: remove weu storage account
+
+## 2.9.4
+
+### Patch Changes
+
+- 6f1df85: remove migration kit for signed documents
+
 ## 2.9.3
 
 ### Patch Changes